requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
web3>=6.0.0
requests==2.31.0
eth-account==0.11.3
eth-hash==0.5.1
eth-keyfile==0.9.1
eth-keys==0.6.1
eth-rlp==1.0.1
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
rlp==4.1.0
